### 2024-12-02 poissonconsulting/bisonpicsuite#11 @aylapear
Fix DOI references in paper bib 
Need to review and update any that need to be fixed.

🟡 SKIP DOIs

- No DOI given, and none found for title: Modelling population dynamics: model formulation, ...
- No DOI given, and none found for title: Statistical rethinking: a Bayesian course with exa...
- No DOI given, and none found for title: Integrated population models: theory and ecologica...

❌ MISSING DOIs

- 10.3389/fevo.2020.563477 may be a valid DOI for title: Spatially Explicit Capture-Recapture Through Camer...

❌ INVALID DOIs

- https://doi.org/10.3106/041.038.0103 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1002/ecs2.3523 is INVALID because of 'https://doi.org/' prefix

https://github.com/openjournals/joss-reviews/issues/7547#issuecomment-2512510236
