+++ paper/paper.bib
@@ -242,7 +242,7 @@ Publisher: Nature Publishing Group},
 	title = {Evaluation of {Camera} {Trap} {Surveys} for {Estimation} of {Sika} {Deer} {Herd} {Composition}},
 	volume = {38},
 	url = {https://bioone.org/journals/mammal-study/volume-38/issue-1/041.038.0103/Evaluation-of-Camera-Trap-Surveys-for-Estimation-of-Sika-Deer/10.3106/041.038.0103.short},
-	doi = {https://doi.org/10.3106/041.038.0103},
+	doi = {10.3106/041.038.0103},
 	number = {1},
 	urldate = {2024-11-26},
 	journal = {Mammal Study},
@@ -256,7 +256,7 @@ Publisher: Nature Publishing Group},
 @article{horneestimating2021,
 	title = {Estimating survival of unmarked neonates with camera traps - {Horne} - 2021 - {Ecosphere} - {Wiley} {Online} {Library}},
 	url = {https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.3523},
-	doi = {https://doi.org/10.1002/ecs2.3523},
+	doi = {10.1002/ecs2.3523},
 	number = {12(6)},
 	urldate = {2024-11-26},
 	journal = {Ecosphere},
