+++ paper/paper.bib
@@ -164,7 +164,8 @@
 	date = {2020},
 	journal = {Frontiers in Ecology and Evolution},
 	volume = {8},
-	url = {https://www.frontiersin.org/articles/10.3389/fevo.2020.563477}
+	url = {https://www.frontiersin.org/articles/10.3389/fevo.2020.563477},
+	doi = {10.3389/fevo.2020.563477}
 }
 
 @article{nakashima2020,
