### 2024-12-02 poissonconsulting/bisonpicsuite#11 @aylapear
- Modelling population dynamics: model formulation, fitting and assessment using state-space methods
   - Paper book so doesn't have a DOI
- Statistical rethinking:
   - this is a course and thus shouldn't have a DOI
- Integrated population models: theory and ecological applications with R and JAGS
   - Paper book so doesn't have DOI
