+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{fuller2007,
-	title = {Reproduction and Survival of Yellowstone Bison},
+	title = {Reproduction and Survival of {Y}ellowstone Bison},
 	author = {Fuller, J. and Garrott, R. and White, P.J. and Aune, K. and Roffe, T. and Rhyan, J.},
 	year = {2007},
 	month = {09},
@@ -12,7 +12,7 @@
 }
 
 @article{wittmer2005,
-	title = {Population dynamics of the endangered mountain ecotype of woodland caribou (Rangifer tarandus caribou) in British Columbia, Canada},
+	title = {Population dynamics of the endangered mountain ecotype of woodland caribou (Rangifer tarandus caribou) in {B}ritish {C}olumbia, {C}anada},
 	author = {Wittmer, Heiko and Mclellan, Bruce and Seip, Dale and Young, James and Kinley, Trevor and Watts, Glen and Hamilton, Dennis},
 	year = {2005},
 	month = {03},
@@ -215,7 +215,7 @@ Publisher: Nature Publishing Group},
 }
 
 @book{mcelreath2016,
-	title = {Statistical rethinking: a Bayesian course with examples in R and Stan},
+	title = {Statistical rethinking: a {B}ayesian course with examples in {R} and {S}tan},
 	author = {McElreath, Richard},
 	year = {2016},
 	date = {2016},
