### 2024-12-03 be8ec21
Reduce word count to == 1000 words
