+++ paper/paper.md
@@ -29,7 +29,7 @@ tags:
 Knowledge of population status and trends is integral to the effective conservation and management of wildlife populations. 
 To this end, demographic ratios (e.g., Calf:Cow ratios, Fig. 1) are routinely used in wildlife management as a readily observable measure of productivity and for forecasting population trajectories [@fuller2007; @wittmer2005; @bender2006]. 
 More complex state-space population modeling approaches [@buckland2004; @paterson2019; @mizuki2020; @newman2023] can also be used to estimate survival and fecundity.
-To understand the age composition of a population, both of these approaches require classified (by age and sex) counts of individuals.
+To understand the age composition of a population, both approaches require classified (by age and sex) counts of individuals.
 
 In remote areas, estimates of herd size and composition are typically obtained from aerial surveys. 
 However, distinguishing animal age and sex can be challenging, particularly in forested environments if animals flee or hide from aircraft. 
@@ -37,18 +37,17 @@ Remote cameras present an alternative method for obtaining classified counts.
 Wildlife cameras have been used to estimate species occupancy, density, behaviour, and to identify individuals through distinct markings [@magoun2011; @steenweg2016; @caravaggi2017; @green2020; @nakashima2020; @singh2022]. 
 Although different age and sex classes can be readily distinguished in remote camera photos for many ungulate species [@laskin2020], it is rare that classified count data are used to estimate population parameters without using individual markings (but see @horneestimating2021, @chitwoodare2017, @ikedaevaluation2013).
 
-![](ratio-plot.png "Figure 1. Data exploration plot from the shinybisonpic app, showing calf:cow ratios over time from individual remote cameras in the Ronald Lake Wood Bison range in northeast Alberta. Ratios are shown with camera trap ID, date of observation, study year, season, and group size. A ratio of 0 represents a group of entirely cows, while an infinite ratio (Inf) represents a group of entirely calves.")
-
-*Figure 1. Data exploration plot from the shinybisonpic app, showing calf:cow ratios over time from individual remote cameras in the Ronald Lake Wood Bison range in northeast Alberta. Ratios are shown with camera trap ID, date of observation, study year, season, and group size. A ratio of 0 represents a group of entirely cows, while an infinite ratio (Inf) represents a group of entirely calves.*
+![*Figure 1. Data exploration plot from the shinybisonpic app, showing calf:cow ratios over time from individual remote cameras in the Ronald Lake Wood Bison range in northeast Alberta. Ratios are shown with camera trap ID, observation date, study year, season, and group size. A ratio of 0 represents a group of entirely cows, while an infinite ratio (Inf) represents a group of entirely calves.* \label{fig:ratioplot}](ratio-plot.png) 
 
 # Summary
 
 We present a method utilizing classified counts from remote cameras to evaluate wood bison (*Bison bison athabascae*) herd demographics.
 We modeled demographic ratios, survival, and productivity using a Bayesian integrated population model (IPM) to combine stage-structured information from multiple data sources and estimate demographic states and transitions [@schaubintegrated2022]. 
-The data included the classified counts from camera trap observations, and census and proportion of calves estimates from aerial surveys.
+The data included classified counts from camera trap observations and census and proportion of calves estimates from aerial surveys.
 
-The counts of classified individuals from each camera trap observation were represented by a series of binomial distributions, as opposed to a multinomial distribution, which allowed us to account for individuals that were classified by age, but not by sex.
-The binomial distributions informed both the counts of each age class (calf, yearling, adult) and corresponding sex ratios.
+
[…]
