### 2024-12-03 a3bc633
- Reduce word count below 1000 words Closes #12
