### 2024-12-03 fd5cbd9
Update paper/paper.bib
Co-authored-by: Nicole Hill <nicole@poissonconsulting.ca>
