+++ paper/paper.bib
@@ -12,7 +12,7 @@
 }
 
 @article{wittmer2005,
-	title = {Population dynamics of the endangered mountain ecotype of woodland caribou (Rangifer tarandus caribou) in {B}ritish {C}olumbia, {C}anada},
+	title = {Population dynamics of the endangered mountain ecotype of woodland caribou ({R}angifer tarandus caribou) in {B}ritish {C}olumbia, {C}anada},
 	author = {Wittmer, Heiko and Mclellan, Bruce and Seip, Dale and Young, James and Kinley, Trevor and Watts, Glen and Hamilton, Dennis},
 	year = {2005},
 	month = {03},
