### 2024-12-03 0b488ee
Update paper/paper.bib
Co-authored-by: Nicole Hill <nicole@poissonconsulting.ca>
