+++ paper/paper.bib
@@ -215,7 +215,7 @@ Publisher: Nature Publishing Group},
 }
 
 @book{mcelreath2016,
-	title = {Statistical rethinking: a {B}ayesian course with examples in {R} and {S}tan},
+	title = {{S}tatistical {R}ethinking: a {B}ayesian course with examples in {R} and {S}tan},
 	author = {McElreath, Richard},
 	year = {2016},
 	date = {2016},
