+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   bisonpic-user-guide: bisonpic-user-guide.html
   bisonpicsuite-getting-started: bisonpicsuite-getting-started.html
-last_built: 2024-12-03T17:18Z
+last_built: 2024-12-03T18:33Z
 urls:
   reference: https://poissonconsulting.github.io/bisonpicsuite/reference
   article: https://poissonconsulting.github.io/bisonpicsuite/articles
