### 2024-12-03 #7547 @aylapear
@kthyng

- We have reduce the paper to be under 1000 words.
- Cleaned up the capitalization in our bib file.

Thanks!
