### 2024-12-03 poissonconsulting/bisonpicsuite#11 @aylapear
Done and resubmitted paper
