### 2024-12-10 #7547 @aylapear
Ok thanks for letting me know. @kthyng Do you have an approximate timeline that the waitlist will take? Will it be a week, a month, 6 months, etc? This is my first time submitting to JOSS so trying to get a general understanding of the process. Thanks
