### 2024-12-10 #7547 @kthyng
@aylapear The wait depends on when relevant editors become available, which I can't predict. But, I would assume it would take a couple of months and then hopefully it will be less than that. People do slow down around the holidays though so that is another factor slowing things down in addition to our regular backlog (we're working on getting more editors now too!).
