### 2025-01-14 c3a53b4
Updated _pkgdown.yml to new Poisson template
