+++ _pkgdown.yml
@@ -1,4 +1,3 @@
 url: https://poissonconsulting.github.io/bisonpicsuite/
 template:
-  bootstrap: 5
-
+  package: poissontemplate
