### 2025-01-14 f058dde
Updated DESCRIPTION to new Poisson template
