+++ DESCRIPTION
@@ -40,5 +40,5 @@ Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
-Config/Needs/website: rmarkdown
+Config/Needs/website: poissonconsulting/poissontemplate
 VignetteBuilder: knitr
