### 2025-01-28 #7547 @mengqi-z
👋 @jedalong @see24 - Would any of you be interested and available to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
