### 2025-02-03 #7547 @mengqi-z
👋 @BrandonEdwards @fernandomayer - Would any of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
