### 2025-02-05 #7547 @BrandonEdwards
> 👋 [@BrandonEdwards](https://github.com/BrandonEdwards) [@fernandomayer](https://github.com/fernandomayer) - Would any of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Yes, happy to review!
