### 2025-02-05 #7752 @editorialbot
[REVIEW]: bisonpicsuite: A set of R packages to estimate wood bison population parameters from remote camera data.
**Submitting author:** @aylapear (<a href="http://orcid.org/0000-0001-7388-1222">Ayla Pearson</a>)
**Repository:** https://github.com/poissonconsulting/bisonpicsuite
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.0
