### 2025-03-28 poissonconsulting/bisonpicsuite#17 @BrandonEdwards
No LICENSE File for JOSS review
Hi! I'm one of the JOSS reviewers for bisonpicsuite openjournals/joss-reviews#7752, one of the requirements is to have a plain-text LICENSE or COPYING file in the repo which is not here. It's looking like (once I finish the rest of the review) that should be the only thing needed for acceptance with JOSS.

Thanks!
