+++ NEWS.md
@@ -1,5 +1,10 @@
 <!-- NEWS.md is maintained by https://fledge.cynkra.com, contributors should not edit this file -->
 
+# bisonpicsuite 0.1.1
+
+- Adding plan text file to the package.
+- Internal updates. 
+
 # bisonpicsuite 0.1.0
 
 - Initial public release.
