### 2025-04-01 poissonconsulting/bisonpicsuite#17 @aylapear
Added licence
