### 2025-05-30 #7752 @mengqi-z
👋 @njtierney - Would you be available to review this submission for JOSS?
