### 2025-06-02 #7752 @mengqi-z
Hi @njtierney - Thank you for your reply! We typically allow reviewers around 4 to 6 weeks to complete a review, so a start date around July 7th should still work. If you're comfortable proceeding with that timing, I’d appreciate your review. Please don’t hesitate to let me know if you'd prefer a bit more flexibility.
