+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   bisonpic-user-guide: bisonpic-user-guide.html
   bisonpicsuite-getting-started: bisonpicsuite-getting-started.html
-last_built: 2025-04-01T17:27Z
+last_built: 2025-06-03T22:06Z
 urls:
   reference: https://poissonconsulting.github.io/bisonpicsuite/reference
   article: https://poissonconsulting.github.io/bisonpicsuite/articles
