### 2025-06-20 poissonconsulting/bisonpicsuite#19 @fernandomayer
JOSS review
Regarding the JOSS review in https://github.com/openjournals/joss-reviews/issues/7752, I just have two concerns regarding these two topics in the software paper:

- Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- State of the field: Do the authors describe how this software compares to other commonly-used packages?

I believe that it would be useful to include in the summary an extended description of the package itself, or better, a brief description of what each of the packages it calls does. I see they're mentioned at the end, but that could be a bit expanded, and perhaps the first figure from [this vignette](https://poissonconsulting.github.io/bisonpicsuite/articles/bisonpicsuite-getting-started.html) could also be included.

Also it would be useful to compare with any other packages. If that doesn't exist, then it worths mention that you found no other package that does the same thing.

And a small correction in the text: "Figure 1" and "Figure 2" are repeated in their respective legends.
