### 2025-06-27 ef48f5e
removing double figure text
