### 2025-06-27 992fe64
adding sentence on comparison to other packages
