### 2025-06-27 b10b0c1
adding overview diagram
