### 2025-06-27 b279bed
expanding on package descriptions
