### 2025-06-28 b5cca63
clean up sentence
