### 2025-07-01 c7b43e9
fix up references
