### 2025-07-01 4e158fa
cleaning up sentence to remove a couple words
