### 2025-07-01 39e66c2
shortening figure caption
