### 2025-07-01 93ce26a
shorten figure caption
