+++ paper/paper.md
@@ -32,7 +32,7 @@ More complex population modeling approaches [@buckland2004; @paterson2019; @mizu
 To understand the age composition of a population, both approaches require classified (by age and sex) counts of individuals.
 
 The use of remote cameras to monitor wildlife species for conservation and management has become widespread and R packages have been developed to manage camera data and estimate species occupancy, density and abundance [@doser_spabundance_2024; @doser_spoccupancy_2022; @moeller_spacentime_2022; @naimi_sdm_2016; @niedballa_camtrapr_2016].
-In animals without individual markings, other population parameters are seldom estimated (but see [@horneestimating2021; @chitwoodare2017; @ikedaevaluation2013]). 
+In animals without individual markings, other population parameters are seldom estimated, although exceptions exist [@horneestimating2021; @chitwoodare2017; @ikedaevaluation2013]. 
 The authors are unaware of any other R packages that use classified counts from camera data to generate estimates of survival and recruitment, even though different age and sex classes can be readily distinguished in remote camera photos for many species [@laskin2020].
 We demonstrate an R package using a Bayesian integrated population model to generate estimates for these vital rates, combined with a Gaussian process regression to account for spatial and temporal correlations.
 
@@ -69,7 +69,7 @@ Collectively, these methods could be applied non-invasively to a wide array of h
 
 This method is implemented using four related R packages.
 The underlying functionality to check, clean, process, model, and visualize data is provided by `bisonpictools`.
-The other two R packages are R shiny apps that provide a user-friendly interface to `bisonpictools`.
+The other two packages are R shiny apps that provide a user-friendly interface to `bisonpictools`.
 Each app starts with an upload data tab that allows users to upload their data and runs a series of checks to ensure the data follows the required format for the next steps.
 The first app is `shinybisonpic`, a web-based app that allows users to explore data by viewing camera locations on a map in one tab and plots of ratios of selected sex-age groups by location and year (e.g., Fig. 1) in another tab. 
 The second app is `runbisonpic`, a locally-run app that allows users with various skill levels to run the model to estimate the abundance by class, total abundance, survival and fecundity rates, and various sex-age ratios (e.g., Fig. 2). 
