### 2025-07-01 42a1585
remove word
