### 2025-07-01 3df59be
clean up bib reference
