+++ paper/paper.bib
@@ -255,7 +255,7 @@ Publisher: Nature Publishing Group},
 
 
 @article{horneestimating2021,
-	title = {Estimating survival of unmarked neonates with camera traps - {Horne} - 2021 - {Ecosphere} - {Wiley} {Online} {Library}},
+	title = {Estimating survival of unmarked neonates with camera traps},
 	url = {https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.3523},
 	doi = {10.1002/ecs2.3523},
 	number = {12(6)},
