### 2025-07-02 558a852
Consistency with "we" and in-text reference to fig. 3.
