### 2025-07-03 poissonconsulting/bisonpicsuite#19 @aylapear
Thanks you for these comments, we have addressed them and updated the paper.
