### 2025-07-04 poissonconsulting/bisonpicsuite#22 @mengqi-z
Joss Paper Proofread
Hi @aylapear,

The paper looks good. I only have one minor comments on the paper for the submission openjournals/joss-reviews#7752.

Please format the app name `runbisonpic` as inline code on line 63 of the PDF version, same format as other software names. Thanks.
