### 2025-07-04 a07b317
fixes #22 update name to be inline code
