### 2025-07-09 #7752 @aylapear
I am unable to tick off the items in your list so copied them to another comment so I can tick them off as I do them

- [ ] Double check authors and affiliations (including ORCIDs).

- [ ] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).

- [ ] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.

- [ ] Please post the DOI of the archived version here.
