### 2025-07-09 8d7d955
fledge: Bump version to 0.1.2
