+++ NEWS.md
@@ -1,5 +1,24 @@
 <!-- NEWS.md is maintained by https://fledge.cynkra.com, contributors should not edit this file -->
 
+# bisonpicsuite 0.1.2
+
+- Merge pull request #23 from poissonconsulting/joss-updates-2.
+
+  Joss updates 2
+
+- Merge pull request #21 from poissonconsulting/joss-updates-2.
+
+  Joss updates 2
+
+- Merge pull request #20 from poissonconsulting/joss-updates-2.
+
+  Joss review comment updates
+
+- Merge pull request #18 from poissonconsulting/joss-update.
+
+  JOSS Review update
+
+
 # bisonpicsuite 0.1.1
 
 - Adding plan text file to the package.
